api/requirements.txt
django
djangorestframework
django-model-utils
gunicorn
psycopg2
factory-boy
pytest
pytest-django